// src/app/modules/investment-approval/interfaces/investment-approval-dashboard-employee-info.ts
export interface InvestmentApprovalDashboardEmployeeInfo {
  employeeMasterId: number;
  employeeCode: string;
  employeeName: string;
  itGroup: string;
  itSection: string;
  type: string;
  proofSubmissionId: string;
  proofSubmissionIdDetail: string;
  dateOfSubmission: Date;
  status: string;
}